sfsfsf

+++ resources/js/house/Bien.vue
@@ -1,3 +1,4 @@
+import Swal from "sweetalert2";
 <template>
     <div id="content">
         <div class="container-fluid">
@@ -10,20 +11,30 @@
                         </div>
                     </div>
                 </div>
-                <div style="z-index: 1000"> <FlashMessage></FlashMessage></div>
+                <div style="z-index: 1000"></div>
 
             </div>
 
 
             <div class="row justify-content-center mt-5">
                 <div class="col-md-11">
+                    <div class="row">
+                        <div class="col-6  ">
+                            <div class="input-group col-md-6 mb-3">
+                                <input type="text" class="form-control"  v-model="Q" placeholder="Recherche">
+
+                            </div>
+                        </div>
+                        <div class="col-6 text-right">
+                            <router-link to="/newBien" class="btn btn-outline-success text-success">Nouveau bien</router-link>
+                        </div>
+                    </div>
                     <div class="text-right mb-3">
-                        <router-link to="/newBien" class="btn btn-outline-success text-success">Nouveau bien</router-link>
                     </div>
                     <div class="card">
                         <div class="card-body">
 
-                            <table class="table table-bordered">
+                            <table class="table table-striped table-bordered" style="width:100%">
                                 <thead class=" text-center">
                                 <tr>
                                     <th scope="col">photo</th>
@@ -38,14 +49,18 @@
                                 <tr>
 
                                     <td>
-                                        <img :src=" avatar(bien.photo) " alt="" width="50">
+                                        <img :src=" avatar(bien.photo) " alt="" class="avatar">
 
                                     </td>
                                     <td>
                                         <router-link :to="{ name: 'showbien', params: { id: bien.id }}" class="text-primary">{{bien.name}}</router-link>
                                     </td>
-                                    <td>{{bien.addresse}}</td>
-                                    <td>{{bien.ville}}</td>
+                                    <td v-if="bien.parentid !== null">{{bien.parentid.addresse}}</td>
+                                    <td v-else>{{bien.addresse}}</td>
+
+                                    <td v-if="bien.parentid !== null">{{bien.parentid.ville}}</td>
+                                    <td v-else>{{bien.ville}}</td>
+
                                     <th class="text-center">
                                         <div class="btn-group">
                                             <button class="btn btn-secondary btn-sm " type="button" data-toggle="dropdown" aria-expanded="false">
@@ -66,6 +81,7 @@
 
                                 </tbody>
                             </table>
+<!--                            <pagination :data="Biens" @pagination-change-page="getResultsPaginate" class="mt-3"></pagination>-->
 
                         </div>
                     </div>
@@ -77,12 +93,15 @@
 </template>
 
 <script>
+    import Swal from 'sweetalert2'
+
     export default {
         name: "Bien.vue",
 
         data(){
             return{
-                Biens:""
+                Biens:{},
+                Q:''
             }
         },
 
@@ -90,7 +109,7 @@
         created(){
             axios.get('api/bien')
                 .then((response)=>{
-                    // console.log(response.data)
+                    console.log(response.data)
                     this.Biens = response.data
                 })
         },
@@ -103,25 +122,44 @@
             },
 
             deleteBien(id){
-                axios.delete('api/bien/' + id)
-                    .then((response)=>{
-                        // console.log(response.data)
+                Swal.fire({
+                    text: "Etes-vous de cette action !",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    cancelButtonText: 'Abandonner',
+                    confirmButtonText: 'Oui, supprimer!'
+                }).then((result) => {
+                    if (result.value) {
+                        axios.delete('api/bien/' + id)
+                            .then((response) => {
 
-                        if (response.data){
-                            this.flashMessage.success({
-                                title: 'Supprimer',
-                                message: 'Action reussit',
-                                time: 3000,
-                                flashMessageStyle: {
-                                    backgroundColor: 'linear-gradient(#e66465, #9198e5)'
-                                }
-                            });
-                        } else {
-                            this.flashMessage.error({title: 'Error Message Title', message: 'xxxxxxxxxx'});
-                        }
-                    })
+                            })
+                        Swal.fire(
+                            'Supprimer',
+                            'success'
+                        )
+                    }
+                })
+            },
+
+            // Our method to GET results from a Laravel endpoint
+            getResultsPaginate(page = 1) {
+                // axios.get('api/bien?page=' + page)
+                //     .then(response => {
+                //         this.Biens = response.data;
+                //     });
             },
         },
+
+        computed:{
+            getFilterBiens() {
+                return this.Biens.filter(bien => {
+                    return bien.indentifiant.toLowerCase().includes(this.Q.toLowerCase())
+                })
+            }
+        }
     }
 </script>
