+++ resources/js/rentale/rental.vue
@@ -75,6 +75,8 @@
 </template>
 
 <script>
+    import Swal from 'sweetalert2'
+
     export default {
         name: "rental",
         data(){
@@ -94,23 +96,26 @@
 
         methods:{
             deleteRent(id){
-                console.log(id)
-                axios.delete('api/rentale/' + id)
-                    .then((response)=>{
+                Swal.fire({
+                    text: "Etes-vous de cette action !",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    cancelButtonText: 'Abandonner',
+                    confirmButtonText: 'Oui, supprimer!'
+                }).then((result) => {
+                    if (result.value) {
+                        axios.delete('api/rentale/' + id)
+                            .then((response) => {
 
-                        if (response.data){
-                            this.flashMessage.success({
-                                title: 'Supprimer',
-                                message: 'Action reussit',
-                                time: 3000,
-                                flashMessageStyle: {
-                                    backgroundColor: 'linear-gradient(#e66465, #9198e5)'
-                                }
-                            });
-                        } else {
-                            this.flashMessage.error({title: 'Error Message Title', message: 'xxxxxxxxxx'});
-                        }
-                    })
+                            })
+                        Swal.fire(
+                            'Supprimer',
+                            'success'
+                        )
+                    }
+                })
             },
         },
