+++ resources/js/tenant/Tenants.vue
@@ -7,7 +7,6 @@
                     <div class="card">
                         <div class="card-body">
                             <div class="float-right ">
-                                <a href="" class="btn btn-outline-secondary">mlkmfkhmfm</a>
 
                             </div>
                         </div>
@@ -20,51 +19,66 @@
             <div class="row justify-content-center mt-5">
                 <div class="col-md-11">
                     <div class="text-right mb-3">
-                        <router-link to="/new-tenant" class="btn btn-outline-success text-success">Nouveau locataire</router-link>
+                        <div class="row">
+                            <div class="col-6  ">
+                                <div class="input-group col-md-6 mb-3">
+                                    <input type="text" class="form-control" v-model="Q" placeholder="Recherche">
+
+                                </div>
+                            </div>
+                            <div class="col-6 text-right">
+                                <router-link to="/new-tenant" class="btn btn-outline-success text-success">Nouveau locataire</router-link>
+                            </div>
+                        </div>
+
                         <FlashMessage class="flashmessage"></FlashMessage>
 
                     </div>
                     <div class="card">
-                        <div class="card-body">
-                            <table class="table table-bordered">
-                                <thead class=" text-center">
+                        <div class="card-body" >
+                            <table id="example" class="table table-striped table-bordered" style="width:100%">
+                                <thead>
                                 <tr>
-                                    <th scope="col">photo</th>
-                                    <th scope="col">Nom</th>
-                                    <th scope="col">Email</th>
-                                    <th scope="col">Telephone</th>
-                                    <th scope="col"> action </th>
+                                    <th><input type="checkbox" onclick="checkAll(this)"></th>
+                                    <th>Profil</th>
+                                    <th>Nom Prenom</th>
+                                    <th>email</th>
+                                    <th>Numero</th>
+                                    <th>Cni</th>
+                                    <th>Actions</th>
                                 </tr>
                                 </thead>
-
-                                <tbody v-for=" tenant in Tenants" v-bind:key="Tenants.id">
+                                <tbody v-for=" tenant in Tenants.data" v-bind:key="Tenants.id">
                                 <tr>
+                                    <td><input type="checkbox" name=""></td>
                                     <td>
-<!--                                        <img :src=" 'image/+tenant.photo ' " alt="" width="50">-->
-                                        <img :src=" avatar(tenant.photo) " alt="" width="50">
+                                        <img :src=" avatar(tenant.photo) " alt="" class="avatar">
                                     </td>
                                     <td>
                                         <router-link :to="{ name: 'showtenant', params: { id: tenant.id }}" class="text-primary">{{tenant.nom}} {{tenant.prenom}}</router-link>
                                     </td>
                                     <td>{{tenant.email}}</td>
                                     <td>{{tenant.numero}}</td>
-                                    <th class="text-center">
+                                    <td>{{tenant.cni}}</td>
+                                    <td>
                                         <div class="btn-group">
-                                            <button class="btn btn-secondary btn-sm " type="button" data-toggle="dropdown" aria-expanded="false">
-                                                actions
-                                            </button>
+                                            <button class="btn btn-secondary btn-sm " type="button" data-toggle="dropdown" aria-expanded="false">actions </button>
                                             <div class="dropdown-menu">
                                                 <router-link :to="{ name: 'edit', params: { id: tenant.id }}" class="dropdown-item"><i class="fas fa-edit fa-sm"> </i> modifier</router-link>
 
                                                 <button @click="deleteTenant(tenant.id)" class="dropdown-item"> <i class="fas fa-trash-alt fa-sm"></i> suprimer</button>
+                                                <vue-confirm-dialog></vue-confirm-dialog>
                                                 <div class="dropdown-divider"></div>
                                                 <a class="dropdown-item" href="#">Autres</a>
                                             </div>
                                         </div>
-                                    </th>
+                                    </td>
                                 </tr>
                                 </tbody>
                             </table>
+
+                            <pagination :data="Tenants" @pagination-change-page="getResultsPaginate" class="mt-3"></pagination>
+
                         </div>
                     </div>
                 </div>
@@ -75,24 +89,26 @@
 </template>
 
 <script>
+    import Swal from 'sweetalert2'
+
+
     export default {
         name: "Tenants",
 
         data(){
             return{
-                Tenants:"",
-
-
-
+                Tenants:{},
+                Q:"",
             }
         },
+
 
         created(){
             axios.get('api/tenants')
                 .then((response)=>{
-                    // console.log(response.data)
+                    console.log(response.data)
                     this.Tenants = response.data
-            })
+                })
         },
 
         methods:{
@@ -103,25 +119,45 @@
             },
 
             deleteTenant(id){
-                axios.delete('api/tenants/' + id)
-                .then((response)=>{
-                    this.tenants = response.data
-                    if (response.data){
-                        this.flashMessage.success({
-                            title: 'Supprimer',
-                            message: 'Action reussit',
-                            time: 3050,
-                            flashMessageStyle: {
-                                backgroundColor: 'linear-gradient(#e66465, #9198e5)',
-                                position:top,
-                            }
-                        });
-                    } else {
-                        this.flashMessage.error({title: 'Error Message Title', message: 'xxxxxxxxxx'});
+                Swal.fire({
+                    text: "Etes-vous de cette action !",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    cancelButtonText: 'Abandonner',
+                    confirmButtonText: 'Oui, supprimer!'
+                }).then((result) => {
+                    if (result.value) {
+                        axios.delete('api/tenants/' + id)
+                            .then((response) => {
+
+                            })
+                        Swal.fire(
+                            'Supprimer',
+                            'success'
+                        )
                     }
                 })
             },
+
+            // Our method to GET results from a Laravel endpoint
+            getResultsPaginate(page = 1) {
+                axios.get('api/tenants?page=' + page)
+                    .then(response => {
+                        this.Tenants = response.data;
+                    });
+            },
+
         },
+
+        computed: {
+            filteredList(){
+                return this.Tenants.filter(tenants => {
+                    return tenants.includes(this.Q.toLowerCase())
+                })
+            }
+        }
     }
 </script>
